move partitioning code
